+++ src/lib/utils/index.ts
@@ -1,4 +1,3 @@
 
 export { Config } from './config/config.contract';
 export { defaultConfig } from './config/config';
-export { Optional } from './optional';