+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added tests to handlers
 - Reworked StorageHandler, it now provides more type-safety.
 - Added storageType to config object and simplified cache config object.
+- Updated dependencies
 
 ## 0.6.0 - Structural upgrade
 - Added Jest unit-testing framework